backend/requirements.txt
fastapi
numpy
opencv-python
phue
pydantic
python-dotenv
scikit-learn
spotipy
uvicorn[standard]